Making the file name "/proc/helloworld" same as others
In this paragraph,  the file name "proc helloworld" in the second sentance is not consistent with that of the latter sentences. The latter sentences use the file name "/proc/helloworld".
Otherwise, I suggest that the  "the file proc helloworld" in the second sentance should change to "the file proc named helloworld".