Provide detailed info for reference count.
This patch is intented to provide detailed info for reference count.

I add sudo lsmod command to provide an alternative way for user to check
the reference count for specific module.

I think `reference count` would be better compared to `use count` so I
replace `use count` with `reference count`.

The paragraph says "there are functions which let you increase,
decrease and display this counter".
But it didn't provide the function to display the the reference counter.
So I add description for module_refcount(THIS_MODULE) as the function
to display the the reference counter.