Update README.md
In line 30, it is sufficient to install xetex in order to save time, and to avoid compiling errors due to multi-version tex installation.

texlive-full includes all kinds of tex, saying luatex, xetex, pdflatex, and the installation takes time. 

Better to install texlive-xetex instead, after that `make` command works which creates the pdf file.